requirements.txt
asposestorage==1.0.2
grad_cam==1.4.6
matplotlib==3.5.2
mpld3==0.5.8
numpy==1.23.0
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==9.4.0
plotly==5.11.0
ptflops==0.6.9
requests==2.27.1
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.8.0
skimage==0.0
submitit==1.4.5
timm==0.6.12
torch==1.9.1+cu111
torchsummary==1.5.1
torchvision==0.10.1+cu111
tqdm==4.64.0
wandb==0.12.21
